revert fn swap
